pyproject.toml
[tool.poetry]
name = "gemini-swim"
version = "0.1.0"
description = ""
authors = ["momonong <morrischen2.9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.*"
asyncpg = "^0.29.0"
databases = "^0.9.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
ipykernel = "^6.29.5"
openpyxl = "^3.1.5"
sqlalchemy = "^2.0.31"
langchain-core = "^0.2.19"
langchain-google-genai = "^1.0.7"
fastapi = "^0.111.1"
uvicorn = "^0.30.1"
dash = "^2.17.1"
flask = {extras = ["async"], version = "^3.0.3"}
flask-caching = "^2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
